+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
+# [2.2.2] - 24-08-2020
+
+### Fixes
+
+- Preview 5 border styling
+
+# [2.2.1] - 21-08-2020
+
+### Fixes
+
+- Downgrade babel-plugin-styled-components to version working properly with Rollup
+
+# [2.2.0] - 20-08-2020
+
+### Added
+
+- Theme background color palette
+
 # [2.1.2] - 10-08-2020
 
 ### Changes
@@ -83,7 +101,6 @@ All notable changes to this project will be documented in this file.
 - Added optional activePeriod prop to the date picker where user defines which presets period to be selected
 - Added CUSTOM period which is used when user chose custom dates from date picker
 - Added style for active date preset
-- Availability to add default or custom presets to date picker and customAllTimeDate
 - Removed margin on Typography
 - Added animation on accordion open/close
 - Do not show the arrow on disabled accordion panel
@@ -561,7 +578,7 @@ All notable changes to this project will be documented in this file.
 
 ### Change
 
-- Availability to add default or custom presets to date picker and customAllTimeDate
+- Availability to add default or custom presets to DatePicker
 
 # [2.0.0-beta.6] - 08-11-2019